src/main/java/br/com/fiap/lanchonete/pedidoservicefase4/domain/entities/Cliente.java
package br.com.fiap.lanchonete.pedidoservicefase4.domain.entities;

public class Cliente {

	private Long id;
	private String nome;
	private String cpf;
	private String telefone;
	private String endereco;
	private String email;

	public Cliente(String nome, String cpf, String telefone, String endereco, String email) {
		this.nome = nome;
		this.cpf = cpf;
		this.telefone = telefone;
		this.endereco = endereco;
		this.email = email;
	}

	public Cliente() {
	}

	public String getNome() {
		return nome;
	}

	public String getCpf() {
		return cpf;
	}

	public String getTelefone() {
		return telefone;
	}

	public String getEndereco() {
		return endereco;
	}

	public String getEmail() {
		return email;
	}

	public void setNome(String nome) {
		this.nome = nome;
	}

	public void setCpf(String cpf) {
		this.cpf = cpf;
	}

	public void setTelefone(String telefone) {
		this.telefone = telefone;
	}

	public void setEndereco(String endereco) {
		this.endereco = endereco;
	}

	public void setEmail(String email) {
		this.email = email;
	}

	public Long getId() {
		return id;
	}

	public void setId(Long id) {
		this.id = id;
	}
	
}
